% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{pkgapi_endpoint}
\alias{pkgapi_endpoint}
\title{Basic endpoint objects}
\description{
Create a \code{pkgapi_endpoint} object that collects
  together an HTTP method (e.g., \code{GET}), a path (e.g.,
  \code{/path}) and a target R function.  Unlike plumber
  endpoints, pkgapi endpoints are meant to be used in testing.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pkgapi_endpoint$new()}}
\item \href{#method-run}{\code{pkgapi_endpoint$run()}}
\item \href{#method-process}{\code{pkgapi_endpoint$process()}}
\item \href{#method-validate_response}{\code{pkgapi_endpoint$validate_response()}}
\item \href{#method-plumber}{\code{pkgapi_endpoint$plumber()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create an endpoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pkgapi_endpoint$new(method, path, target, validate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{The HTTP method to support}

\item{\code{path}}{The server path for the endpoint}

\item{\code{target}}{An R function to run as the endpoint}

\item{\code{validate}}{Logical, indicating if any validation
(implemented by a subclass) should be enabled.  This should be
set to \code{FALSE} in production environments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
Run the endpoint.  This will produce a
standardised response pbject that contains \code{status_code},
\code{content_type}, \code{body} (the serialised output as tun
through the \code{process} method and returned by plumber) and
\code{data} (the result of running the target function)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pkgapi_endpoint$run(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed through to the \code{target} function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process"></a>}}
\subsection{Method \code{process()}}{
Process the data (result of \code{target}),
including all serialisation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pkgapi_endpoint$process(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data to process (the result of \code{target})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_response"></a>}}
\subsection{Method \code{validate_response()}}{
Validate that the body conforms to the schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pkgapi_endpoint$validate_response(body)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{body}}{The serialised body as returned by \code{process}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plumber"></a>}}
\subsection{Method \code{plumber()}}{
Helper method for use with plumber - not designed
for end-user use.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pkgapi_endpoint$plumber(req, res, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req, res}}{Conventional plumber request/response objects}

\item{\code{...}}{Additional arguments passed through to \code{run}}
}
\if{html}{\out{</div>}}
}
}
}
