% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{pkgapi_endpoint_binary}
\alias{pkgapi_endpoint_binary}
\title{Create a binary endpoint}
\usage{
pkgapi_endpoint_binary(method, path, target, validate = FALSE)
}
\arguments{
\item{method}{The HTTP method to support}

\item{path}{The server path for the endpoint}

\item{target}{An R function to run as the endpoint}

\item{validate}{Logical, indicating if any validation (implemented
by the \code{validate_response} argument) should be enabled.
This should be set to \code{FALSE} in production environments.}
}
\description{
Create a binary endpoint
}
