% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{pkgapi_endpoint_json}
\alias{pkgapi_endpoint_json}
\title{Create JSON endpoint}
\usage{
pkgapi_endpoint_json(
  method,
  path,
  target,
  validate = FALSE,
  schema = NULL,
  root = NULL
)
}
\arguments{
\item{method}{The HTTP method to support}

\item{path}{The server path for the endpoint}

\item{target}{An R function to run as the endpoint}

\item{validate}{Logical, indicating if any validation (implemented
by the \code{validate_response} argument) should be enabled.
This should be set to \code{FALSE} in production environments.}

\item{schema}{The name of the json schema to use}

\item{root}{The root of the schema directory.  If not provided,
and if \code{target} is in a package, then we'll look in that
package's installed \code{schema} directory.}
}
\description{
Create a JSON endpoint
}
\details{
Note: We look for schema files in \code{<root>/<schema>.json}.
}
