% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{pkgapi_stop}
\alias{pkgapi_stop}
\title{Throw an error from an endpoint}
\usage{
pkgapi_stop(message, code = "ERROR", errors = NULL, status_code = 400L)
}
\arguments{
\item{message}{The human-readable message of the error.  Ignored
if \code{errors} is given.}

\item{code}{Optional code for the error - if not given, then
\code{ERROR} is used.  Ignored if \code{errors} is given.}

\item{errors}{A named list of errors - use this to signal multiple
error conditions as key/value pairs}

\item{status_code}{The HTTP status code to use.  The default (400)
means "bad request" which should be a reasonable catch-all for
bad user data.}
}
\value{
Nothing, as this function throws an error
}
\description{
Throw an error from an endpoint.  This function is intended to
allow target functions to throw nice errors back through the API.
}
