% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porcelain.R
\name{porcelain}
\alias{porcelain}
\title{A \code{porcelain} object}
\description{
A \code{porcelain} object.  This extends (via
inheritance) a plumber object, and so only changes to the
plumber API are documented here.
}
\section{Super classes}{
\code{\link[plumber:Hookable]{plumber::Hookable}} -> \code{\link[plumber:Plumber]{plumber::Plumber}} -> \code{porcelain}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{porcelain$new()}}
\item \href{#method-handle_package}{\code{porcelain$handle_package()}}
\item \href{#method-handle}{\code{porcelain$handle()}}
\item \href{#method-request}{\code{porcelain$request()}}
\item \href{#method-clone}{\code{porcelain$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Hookable" data-id="registerHooks">}\href{../../plumber/html/Hookable.html#method-registerHooks}{\code{plumber::Hookable$registerHooks()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="addAssets">}\href{../../plumber/html/Plumber.html#method-addAssets}{\code{plumber::Plumber$addAssets()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="addEndpoint">}\href{../../plumber/html/Plumber.html#method-addEndpoint}{\code{plumber::Plumber$addEndpoint()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="addFilter">}\href{../../plumber/html/Plumber.html#method-addFilter}{\code{plumber::Plumber$addFilter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="addGlobalProcessor">}\href{../../plumber/html/Plumber.html#method-addGlobalProcessor}{\code{plumber::Plumber$addGlobalProcessor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="call">}\href{../../plumber/html/Plumber.html#method-call}{\code{plumber::Plumber$call()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="filter">}\href{../../plumber/html/Plumber.html#method-filter}{\code{plumber::Plumber$filter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="getApiSpec">}\href{../../plumber/html/Plumber.html#method-getApiSpec}{\code{plumber::Plumber$getApiSpec()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="getDebug">}\href{../../plumber/html/Plumber.html#method-getDebug}{\code{plumber::Plumber$getDebug()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="mount">}\href{../../plumber/html/Plumber.html#method-mount}{\code{plumber::Plumber$mount()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="onHeaders">}\href{../../plumber/html/Plumber.html#method-onHeaders}{\code{plumber::Plumber$onHeaders()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="onWSOpen">}\href{../../plumber/html/Plumber.html#method-onWSOpen}{\code{plumber::Plumber$onWSOpen()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="openAPIFile">}\href{../../plumber/html/Plumber.html#method-openAPIFile}{\code{plumber::Plumber$openAPIFile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="print">}\href{../../plumber/html/Plumber.html#method-print}{\code{plumber::Plumber$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="registerHook">}\href{../../plumber/html/Plumber.html#method-registerHook}{\code{plumber::Plumber$registerHook()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="removeHandle">}\href{../../plumber/html/Plumber.html#method-removeHandle}{\code{plumber::Plumber$removeHandle()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="route">}\href{../../plumber/html/Plumber.html#method-route}{\code{plumber::Plumber$route()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="run">}\href{../../plumber/html/Plumber.html#method-run}{\code{plumber::Plumber$run()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="serve">}\href{../../plumber/html/Plumber.html#method-serve}{\code{plumber::Plumber$serve()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="set404Handler">}\href{../../plumber/html/Plumber.html#method-set404Handler}{\code{plumber::Plumber$set404Handler()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="setApiSpec">}\href{../../plumber/html/Plumber.html#method-setApiSpec}{\code{plumber::Plumber$setApiSpec()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="setDebug">}\href{../../plumber/html/Plumber.html#method-setDebug}{\code{plumber::Plumber$setDebug()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="setDocs">}\href{../../plumber/html/Plumber.html#method-setDocs}{\code{plumber::Plumber$setDocs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="setDocsCallback">}\href{../../plumber/html/Plumber.html#method-setDocsCallback}{\code{plumber::Plumber$setDocsCallback()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="setErrorHandler">}\href{../../plumber/html/Plumber.html#method-setErrorHandler}{\code{plumber::Plumber$setErrorHandler()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="setParsers">}\href{../../plumber/html/Plumber.html#method-setParsers}{\code{plumber::Plumber$setParsers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="setSerializer">}\href{../../plumber/html/Plumber.html#method-setSerializer}{\code{plumber::Plumber$setSerializer()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="swaggerFile">}\href{../../plumber/html/Plumber.html#method-swaggerFile}{\code{plumber::Plumber$swaggerFile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="plumber" data-topic="Plumber" data-id="unmount">}\href{../../plumber/html/Plumber.html#method-unmount}{\code{plumber::Plumber$unmount()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a porcelain object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{porcelain$new(..., validate = FALSE, logger = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Parameters passed to \code{\link{plumber}}}

\item{\code{validate}}{Logical, indicating if any validation
(implemented by the \code{validate_response} argument) should
be enabled.  This should be set to \code{FALSE} in production
environments.  By default (if \code{validate} is \code{NULL}),
we look at the value of the environment \code{PORCELAIN_VALIDATE} -
if \code{true} (case insensitive) then we will validate.
This is intended to support easy use of validation on
continuous integration systems.}

\item{\code{logger}}{Optional logger, from the \code{lgr} package.  If
given, then we will log at the beginning and end of the request.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-handle_package"></a>}}
\if{latex}{\out{\hypertarget{method-handle_package}{}}}
\subsection{Method \code{handle_package()}}{
Handle package endpoints
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{porcelain$handle_package(state = NULL, package = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state}}{A named list of state, if your package requires
any state-binding endpoints. Typically these will be mutable
state (database connections, job queues, or similar).  You must
provide all states as required by the combination of all
endpoints.}

\item{\code{package}}{Either a package name or environment (optional,
usually we'll find the right thing)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-handle"></a>}}
\if{latex}{\out{\hypertarget{method-handle}{}}}
\subsection{Method \code{handle()}}{
Handle an endpoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{porcelain$handle(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Either a single argument, being a
\code{\link{porcelain_endpoint}} object representing an endpoint, or
arguments to pass through to \code{plumber}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-request"></a>}}
\if{latex}{\out{\hypertarget{method-request}{}}}
\subsection{Method \code{request()}}{
Send a request to plumber for debugging

Sends a request to plumber so that the API can be easily
tested without running the whole API. The interface here will
probably change, and may end up using the interface of \code{httr}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{porcelain$request(method, path, query = NULL, body = NULL, content_type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of HTTP method to use (e.g., \code{GET})}

\item{\code{path}}{Path to send the request to}

\item{\code{query}}{Optional query parameters as a named list or
character vector.}

\item{\code{body}}{Optional body (only valid with \code{PUT}, \code{POST},
etc).}

\item{\code{content_type}}{Optional content type (mime) which can be
provided alongside \code{body}.  If not provided it is set to
\code{application/octet-stream} if \code{body} is raw, or
\code{application/json} otherwise.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{porcelain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
