% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{porcelain_add_headers}
\alias{porcelain_add_headers}
\title{Add headers to endpoint output data}
\usage{
porcelain_add_headers(data, headers)
}
\arguments{
\item{data}{Response data}

\item{headers}{Named list of headers to add.}
}
\value{
Data from endpoint target with headers
}
\description{
Intended to be used from endpoint target function. Note \code{Content-Type}
headers are handled by returning arg to endpoint.
}
\examples{
porcelain_add_headers("output",
                   list("Content-Dispotition" = "output_file.txt"))
}
