% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{porcelain_logger}
\alias{porcelain_logger}
\title{Create logger}
\usage{
porcelain_logger(log_level = "info", name = NULL, path = NULL)
}
\arguments{
\item{log_level}{The level of detail to log to. See
\code{\link[lgr:get_log_levels]{lgr::get_log_levels()}} for possible values; this could be a
string ("off", "info", "all", etc) or an integer level.}

\item{name}{The name of the logger.  By default we use one derived
from the package name, though this may not always be accurate.}

\item{path}{Optionally, the path to log into.  If not given then
we log to the console.}
}
\value{
A "Logger" object (see \link[lgr:Logger]{lgr::Logger})
}
\description{
Create a json-emitting logger, using the 'lgr' package.
}
\examples{
logger <- porcelain::porcelain_logger(name = "example")
logger$log("info", "hello")
logger$log("trace", "silent")
}
