% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{porcelain_package_endpoint}
\alias{porcelain_package_endpoint}
\title{Find roxygen-defined endpoint}
\usage{
porcelain_package_endpoint(
  package,
  method,
  path,
  state = NULL,
  validate = NULL
)
}
\arguments{
\item{package}{The name of the package to look in, provided as a
string or as a namespace}

\item{method}{The HTTP method (i.e., verb), such as \code{GET} or
\code{POST}, as a string}

\item{path}{The path of the method (e.g., \verb{/my/path})}

\item{state}{A list of state to bind into the method, if your
endpoint requires any}

\item{validate}{Logical, indicating if the method should be
created with schema validation enabled.}
}
\value{
The endpoint, a \link{porcelain_endpoint} object
}
\description{
Find an endpoint defined implicitly through roxygen comments
(rather than explicitly via writing
\link{porcelain_endpoint}.
}
