% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{porcelain_roclet}
\alias{porcelain_roclet}
\title{Define API using roxygen tags}
\usage{
porcelain_roclet()
}
\value{
A roclet, used by \code{roxygen2} (not typically called by
users directly)
}
\description{
A roclet for processing \verb{@porcelain} tags within a package. This
presents an automated declarative approach to defining porcelain
APIs using roxygen tags. When you roxygenise your package (e.g.,
with \code{devtools::document()} or \code{roxygen2::roxygenise()}) this
roclet will create a file \code{R/porcelain.R} within your package that
will be included into your package API.
}
